+++ test/test_jet.jl
@@ -5,6 +5,6 @@
 
     rep = JET.report_package(QuantumSavory; target_modules=(QuantumSavory,))
     println(rep)
-    @test length(JET.get_reports(rep)) <= 10
     @test length(JET.get_reports(rep)) == 0
+    #@test_broken length(JET.get_reports(rep)) == 0
 end